Move BeamSearchSynthesizer to synthesizer
