+++ configs/hearthstone/treegen_evaluate.gin
@@ -52,23 +52,23 @@ nl2prog.utils.data.treegen.CollateQuery.device          = @torch.device()
 nl2prog.utils.data.CollateNlFeature.device              = @torch.device()
 
 # Synthesizer
-nl2prog.utils.CommonBeamSearchSynthesizer.create.beam_size               = 15
-nl2prog.utils.CommonBeamSearchSynthesizer.create.transform_input         = @nl2prog.utils.transform.treegen.TransformQuery()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.transform_evaluator     = @nl2prog.utils.transform.treegen.TransformEvaluator()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.collate_input           = @nl2prog.utils.data.treegen.CollateInput()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.collate_action_sequence = @nl2prog.utils.data.treegen.CollateActionSequence()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.collate_query           = @nl2prog.utils.data.treegen.CollateQuery()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.collate_state           = @nl2prog.utils.data.collate_none
-nl2prog.utils.CommonBeamSearchSynthesizer.create.collate_nl_feature      = @nl2prog.utils.data.CollateNlFeature()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.collate_other_feature   = @nl2prog.utils.data.collate_none
-nl2prog.utils.CommonBeamSearchSynthesizer.create.split_states            = @nl2prog.utils.data.split_none
-nl2prog.utils.CommonBeamSearchSynthesizer.create.model                   = @nl2prog.nn.treegen.TrainModel()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.action_sequence_encoder = @action_sequence_encoder/nl2prog.gin.workspace.get()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.is_subtype              = @nl2prog.language.python.is_subtype
-nl2prog.utils.CommonBeamSearchSynthesizer.create.options                 = @nl2prog.ast.action.ActionOptions()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.device                  = @torch.device()
-nl2prog.utils.CommonBeamSearchSynthesizer.create.max_steps               = 350
-synthesizer/nl2prog.gin.workspace.put.value                              = @nl2prog.utils.CommonBeamSearchSynthesizer.create()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.beam_size               = 15
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.transform_input         = @nl2prog.utils.transform.treegen.TransformQuery()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.transform_evaluator     = @nl2prog.utils.transform.treegen.TransformEvaluator()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.collate_input           = @nl2prog.utils.data.treegen.CollateInput()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.collate_action_sequence = @nl2prog.utils.data.treegen.CollateActionSequence()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.collate_query           = @nl2prog.utils.data.treegen.CollateQuery()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.collate_state           = @nl2prog.utils.data.collate_none
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.collate_nl_feature      = @nl2prog.utils.data.CollateNlFeature()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.collate_other_feature   = @nl2prog.utils.data.collate_none
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.split_states            = @nl2prog.utils.data.split_none
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.model                   = @nl2prog.nn.treegen.TrainModel()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.action_sequence_encoder = @action_sequence_encoder/nl2prog.gin.workspace.get()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.is_subtype              = @nl2prog.language.python.is_subtype
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.options                 = @nl2prog.ast.action.ActionOptions()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.device                  = @torch.device()
+nl2prog.synthesizer.CommonBeamSearchSynthesizer.create.max_steps               = 350
+synthesizer/nl2prog.gin.workspace.put.value                              = @nl2prog.synthesizer.CommonBeamSearchSynthesizer.create()
 
 # Metrics
 nl2prog.metrics.Accuracy.parse          = @nl2prog.language.python.parse
