+++ nl2prog/synthesizer/common_beam_search_synthesizer.py
@@ -7,8 +7,8 @@ from dataclasses import dataclass
 from nl2prog.ast.action import ActionOptions, Rule, CloseNode
 from nl2prog.ast.evaluator import Evaluator
 from nl2prog.encoders import ActionSequenceEncoder
-from nl2prog.utils.beam_search_synthesizer import Hypothesis
-from nl2prog.utils \
+from nl2prog.synthesizer.beam_search_synthesizer import Hypothesis
+from nl2prog.synthesizer \
     import BeamSearchSynthesizer as BaseBeamSearchSynthesizer, \
     IsSubtype, LazyLogProbability
 from nl2prog.nn import TrainModel
