+++ nl2prog/synthesizer/synthesizer.py
@@ -1,0 +1,31 @@
+from typing import Any, Tuple, List, Generator, Dict, Optional
+from nl2prog.ast.action import Action
+from nl2prog.ast.ast import AST
+from dataclasses import dataclass
+
+
+@dataclass
+class Progress:
+    id: int
+    parent: Optional[int]
+    score: float
+    action: Action
+    is_complete: bool
+
+
+@dataclass
+class Candidate:
+    score: float
+    ast: AST
+
+
+class Synthesizer:
+    def synthesize(self, input: Any) \
+            -> Generator[Tuple[List[Candidate], List[Progress]], None, None]:
+        raise NotImplementedError
+
+    def load_state_dict(self, state_dict: Dict[str, Any]) -> None:
+        raise NotImplementedError
+
+    def state_dict(self) -> Dict[str, Any]:
+        raise NotImplementedError
