+++ test/test_nl2code.py
@@ -9,7 +9,8 @@ import os
 import torch
 import torch.optim as optim
 from nl2prog.gin import nl2prog, nl2code, workspace, optimizer
-from nl2prog.utils import Query, CommonBeamSearchSynthesizer
+from nl2prog.utils import Query
+from nl2prog.synthesizer import CommonBeamSearchSynthesizer
 from nl2prog.ast.action \
     import ast_to_action_sequence as to_seq, ActionOptions
 from nl2prog.utils.data \
