+++ test/test_treegen.py
@@ -10,7 +10,8 @@ import torch
 import fairseq.optim as optim
 
 from nl2prog.gin import nl2prog, treegen, optimizer, workspace
-from nl2prog.utils import Query, CommonBeamSearchSynthesizer
+from nl2prog.utils import Query
+from nl2prog.synthesizer import CommonBeamSearchSynthesizer
 from nl2prog.ast.action \
     import ast_to_action_sequence as to_seq, ActionOptions
 from nl2prog.utils.data \
