+++ test/utils/test_synthesize.py
@@ -2,7 +2,7 @@ import unittest
 import ast
 from typing import List
 from nl2prog.language.python import to_ast
-from nl2prog.utils import Progress, Candidate, synthesize as _synthesize
+from nl2prog.synthesizer import Progress, Candidate, synthesize as _synthesize
 
 
 class TestSynthesize(unittest.TestCase):
